# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for A3S. This service allows authentication, authorisation and accounting.
  version: 1.1.3
  title: A3S
paths:
...
  '/subrealms':
    get:
      tags:    
        - SubRealm
      description: Search for Sub-Realms.
      summary: Search for Sub-Realms.
      operationId: listSubRealms
      parameters:
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 1
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: includeRelations
          in: query
          required: false
          description: |
            Determines whether related entities, such as any associated permissions, are returned.
          schema:
            type: boolean
            default: false
        - name: filterName
          in: query
          required: false
          description: A search query filter on the name of the sub-realm.
          schema:
            type: string
            minLength: 0
            maxLength: 255
          example: FilterByEntityType
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append '_desc' after a field to indicate descending order.  Supported fields. 'name'.
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubRealm'

        '204':
          description: No Content.
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of Sub Realms.
          content:
            application/json:
              schema:    
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sub-Realm list not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - SubRealm
      summary: Create a new Sub-Realm.
      description: Create a new Sub-Realm.
      operationId: createSubRealm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubRealmSubmit'
      responses:
        '200':    
          description: Successful. Sub-Realm created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubRealm'
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not authorized to create a Sub-Realm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Non-Processible request. The request was correctly structured but some business rules were violated, preventing the creation of the Sub-Realm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subrealms/{subRealmId}':
    get:
      tags:
        - SubRealm
      summary: Get a Sub-Realm.
      description: Get a Sub-Realm by its UUID.
      operationId: getSubRealm
      parameters:
        - name: subRealmId
          in: path
          description: The UUID of the Sub-Realm.
          required: true
          schema:
            $ref: '#/components/schemas/SubRealmId'
      responses:
        '200':    
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubRealm'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to read Sub-Realms.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sub-Realm not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - SubRealm
      summary: Update a Sub-Realm.
      description: Update a Sub-Realm by its UUID.
      operationId: updateSubRealm
      parameters:
        - name: subRealmId
          in: path
          description: The UUID of the Sub-Realm.
          required: true
          schema:
            $ref: '#/components/schemas/SubRealmId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubRealmSubmit'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubRealm'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:    
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to update Sub-Realms.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sub-Realm not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Non-Processible request. The request was correctly structured but some business rules were violated, preventing the update of the Sub-Realm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - SubRealm
      summary: Deletes a sub-realm.
      description: Deletes a sub-realm.
      operationId: deleteSubRealm
      parameters:
        - name: subRealmId
          in: path
          description: The UUID of the Sub-Realm.
          required: true
          schema:
            $ref: '#/components/schemas/SubRealmId'
      responses:
        '204':
          description: No Content.
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:    
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to delete Sub-Realms.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sub-Realm not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Non-Processible request. Usually returned when the request is correctly structured but some business rules have been violated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  # End Region - Sub Realms
components:
  schemas:
...
    PermissionId:
      type: string
      format: uuid
    Permission:
      description: |
        Permission for an application
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/PermissionId'
        name:
          type: string
        description:
          type: string

    # End Region - Permissions      

    # Region - TermsOfService
...
    ApplicationDataPolicy:
      description: |
          Models an applications data policy.
      type: object
      required:
        - name
        - description
        - uuid
      properties:
        uuid:
          $ref: '#/components/schemas/ApplicationDataPolicyId'
        name:
          type: string
          description: The name of the application data policy.
        description:
          type: string
          description: A description of the application data policy.
    ApplicationDataPolicyId:
      type: string
      format: uuid
    # End Regions - Application Data Policies
    
    # Region - Sub-Realms
    SubRealmId:
      description: The UUID identifier for a sub-realm.
      type: string
      format: uuid
    SubRealm:
      description: |
          Models a sub-realm.
      type: object
      required:
        - name
        - description
        - uuid
        - permissions
        - applicationDataPolicies
      properties:
        uuid:
          $ref: '#/components/schemas/SubRealmId'
        name:
          type: string
          description: The name of the sub-realm.
        description:
          type: string
          description: A description of the sub-realm.
        permissions:
          description: A list of all the permissions that are available within the sub-realm.
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        applicationDataPolicies:
          description: A list of all the application data policies that are available within the sub-realm.
          type: array
          items:
            $ref: '#/components/schemas/ApplicationDataPolicy'
    SubRealmSubmit:
      description: |
          Models a sub-realm.
      type: object
      required:
        - name
        - description
        - permissionIds
        - applicationDataPolicyIds
      properties:
        name:
          type: string
          description: The name of the sub-realm.
        description:
          type: string
          description: A description of the sub-realm.
        permissionIds:
          description: A list of all the permission IDs for the permissions that are to be available within the sub-realm.
          type: array
          items:
            $ref: '#/components/schemas/PermissionId'
        applicationDataPolicyIds:
          description: A list of all the application data policy IDs for the application data policies that are to be available within the sub-realm.
          type: array
          items:
            $ref: '#/components/schemas/ApplicationDataPolicyId'

    #region Responses
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PaginationHeaderResponse:
      description: >-
        This is the standard response for requests that have paginated
        collections
      type: object
      properties:
        total:
          type: integer
          format: int32
          minimum: 1
          description: The total number of pages in the result set
        count:
          type: integer
          format: int32
          minimum: 1
          description: The total number of results in the result set
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          default: 10
          description: >-
            The number items to include in a page of results. The page size for
            the result set
        current:
          type: integer
          format: int32
          minimum: 1
          description: >-
            The position of the page in the paged result set that is being
            returned
        first:
          type: string
          format: url
          description: >-
            The link to the first page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters.
        last:
          type: string
          format: url
          description: >-
            The link to the last page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters.
        prev:
          type: string
          format: url
          description: >-
            The link to the previous page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        next:
          type: string
          format: url
          description: >-
            The link to the next page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the last page in the result set then this will
            be [null].
...